+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added ErrorReport when sending a ping request fails
 - Sos-report's tar command no longer use --verbose
+- ErrorReports list now works with paging and a default limit.
+  Fixing node timeouts on large error-reports
 
 ### Fixed